Technical drawing
This is NOT the technical drawing for this keyboard. This is the technical drawing for the default bakeneko, from which this keyboard is based off of.
This information is important because of the threaded inserts. Factories will want to know the specs for threaded holes (where the screws go).

The M2 in this drawing is what is used for both the daughterboard holes AND the back weight holes.